+++ clips/templates/clips/clip_update_form.html
@@ -1,4 +1,4 @@
-<form action="" method="post">{% csrf_token %}
+<form action="{% url 'clips-update' object.pk %}" method="post">{% csrf_token %}
   {{ form.as_p }}
   <input type="submit" value="Update" />
 </form>
@@ -14,3 +14,8 @@
   {% endfor %}
 </tr>{% endfor %}
 </table>
+
+<form action="{% url 'clips-delete' object.pk %}" method="post">
+  {% csrf_token %}
+  <input type="submit" value="Hide this query" />
+</form>
