+++ requirements/app.txt
@@ -46,3 +46,4 @@ whitenoise==5.0.1
 celery==4.4.3
 amqp==2.5.2
 newrelic==5.14.0.142
+django-safedelete==0.5.5
